refactor(frontend): split media/recommendations/plays/settings/connectors api mixins
Move the streaming/media-url, recommendations, playback-history, settings,
and service-connector domain methods out of api.ts into typed mixin modules
under frontend/lib/api/. Pure structural move.

Refactor slice P34 (api-client-typing-split).
