+++ frontend/lib/api.ts
@@ -3,11 +3,14 @@ import {
     IMPORT_PREVIEW_TIMEOUT_MS,
     toSearchParams,
     type ApiData,
-    type ServiceTestResult,
 } from "./api/core";
+import { WithConnectors } from "./api/connectors";
 import { WithLibrary } from "./api/library";
+import { WithMedia } from "./api/media";
 import { WithPlaylists } from "./api/playlists";
-import type { AddToPlaylistRef } from "./trackRef";
+import { WithPlays } from "./api/plays";
+import { WithRecommendations } from "./api/recommendations";
+import { WithSettings } from "./api/settings";
 import type {
     YouTubePlaylistInfo,
     YouTubeDownloadJob,
@@ -357,7 +360,9 @@ export interface PlaybackClientMetricInput {
     fields?: Record<string, unknown>;
 }
 
-class ApiClient extends WithPlaylists(WithLibrary(ApiClientCore)) {
+class ApiClient extends WithConnectors(WithSettings(WithPlays(WithRecommendations(WithMedia(
+    WithPlaylists(WithLibrary(ApiClientCore))
+))))) {
     // Onboarding
     async getOnboardingStatus(): Promise<{
         needsOnboarding: boolean;
@@ -518,418 +523,10 @@ class ApiClient extends WithPlaylists(WithLibrary(ApiClientCore)) {
         });
     }
 
-    // Streaming
-    getStreamUrl(trackId: string): string {
-        const baseUrl = `${this.getBaseUrl()}/api/library/tracks/${trackId}/stream`;
-        // For audio element requests, cookies may not be sent cross-origin in development
-        // Add token as query param for authentication (supported by requireAuthOrToken)
-        const token = this.getCurrentToken();
-        if (token) {
-            return `${baseUrl}?token=${encodeURIComponent(token)}`;
-        }
-        return baseUrl;
-    }
-
-    async createSegmentedStreamingSession(
-        input: CreateSegmentedStreamingSessionInput,
-    ): Promise<SegmentedStreamingSessionResponse> {
-        const headers: Record<string, string> = {};
-        if (typeof input.startupLoadId === "number" && Number.isFinite(input.startupLoadId)) {
-            headers["x-segmented-startup-load-id"] = String(input.startupLoadId);
-        }
-        if (
-            typeof input.startupCorrelationId === "string" &&
-            input.startupCorrelationId.trim().length > 0
-        ) {
-            headers["x-segmented-startup-correlation-id"] =
-                input.startupCorrelationId.trim();
-        }
-        const session = await this.request<SegmentedStreamingSessionResponse>(
-            "/streaming/v1/sessions",
-            {
-                method: "POST",
-                body: JSON.stringify({
-                    trackId: input.trackId,
-                    sourceType: input.sourceType,
-                    desiredQuality: input.desiredQuality,
-                    manifestProfile: input.manifestProfile,
-                }),
-                headers,
-            },
-        );
-
-        return {
-            ...session,
-            manifestUrl: this.toAbsoluteApiUrl(session.manifestUrl),
-        };
-    }
-
-    getStreamingAuthToken(): string | null {
-        return this.getCurrentToken();
-    }
-
-    async heartbeatSegmentedStreamingSession(
-        sessionId: string,
-        sessionToken: string,
-        snapshot: SegmentedStreamingSnapshotInput,
-    ): Promise<SegmentedStreamingHeartbeatResponse> {
-        return this.request<SegmentedStreamingHeartbeatResponse>(
-            `/streaming/v1/sessions/${encodeURIComponent(sessionId)}/heartbeat`,
-            {
-                method: "POST",
-                body: JSON.stringify(snapshot),
-                headers: {
-                    "x-streaming-session-token": sessionToken,
-                },
-            },
-        );
-    }
-
-    async handoffSegmentedStreamingSession(
-        sessionId: string,
-        sessionToken: string,
-        snapshot: SegmentedStreamingSnapshotInput,
-    ): Promise<SegmentedStreamingHandoffResponse> {
-        const handoff = await this.request<SegmentedStreamingHandoffResponse>(
-            `/streaming/v1/sessions/${encodeURIComponent(sessionId)}/handoff`,
-            {
-                method: "POST",
-                body: JSON.stringify(snapshot),
-                headers: {
-                    "x-streaming-session-token": sessionToken,
-                },
-            },
-        );
-
-        return {
-            ...handoff,
-            manifestUrl: this.toAbsoluteApiUrl(handoff.manifestUrl),
-        };
-    }
-
-    async reportPlaybackClientMetric(
-        input: PlaybackClientMetricInput,
-    ): Promise<void> {
-        await this.request<void>("/streaming/v1/client-metrics", {
-            method: "POST",
-            body: JSON.stringify(input),
-        });
-    }
-
-    /**
-     * Get the URL for cover art.
-     * @param coverId - The cover ID, URL, or path
-     * @param size - Optional size in pixels
-     * @param includeToken - Include auth token in URL (needed for canvas color extraction)
-     */
-    getCoverArtUrl(coverId: string, size?: number, includeToken = true): string {
-        const baseUrl = this.getBaseUrl();
-        const token = includeToken ? this.getCurrentToken() : null;
-
-        // Check if this is an audiobook cover path (served by audiobooks endpoint, not proxied)
-        if (coverId && coverId.startsWith("/audiobooks/")) {
-            const url = `${baseUrl}/api${coverId}`;
-            if (token) {
-                return `${url}?token=${encodeURIComponent(token)}`;
-            }
-            return url;
-        }
-
-        // Check if this is a podcast cover path (served by podcasts endpoint, not proxied)
-        if (coverId && coverId.startsWith("/podcasts/")) {
-            const url = `${baseUrl}/api${coverId}`;
-            if (token) {
-                return `${url}?token=${encodeURIComponent(token)}`;
-            }
-            return url;
-        }
-
-        // Check if coverId is an external URL (needs to be proxied)
-        // Also handle native: paths which need URL encoding
-        if (
-            coverId &&
-            (coverId.startsWith("http://") ||
-                coverId.startsWith("https://") ||
-                coverId.startsWith("native:"))
-        ) {
-            // Pass as query parameter to avoid URL encoding issues
-            const params = new URLSearchParams({ url: coverId });
-            if (size) params.append("size", size.toString());
-            if (token) params.append("token", token);
-            return `${baseUrl}/api/library/cover-art?${params.toString()}`;
-        }
-
-        // Otherwise use as path parameter (cover ID - typically a hash)
-        const params = new URLSearchParams();
-        if (size) params.append("size", size.toString());
-        if (token) params.append("token", token);
-        const queryString = params.toString();
-        return `${baseUrl}/api/library/cover-art/${encodeURIComponent(coverId)}${
-            queryString ? "?" + queryString : ""
-        }`;
-    }
-
-    /**
-     * Get the proxied URL for a YouTube Music browse thumbnail.
-     * @param externalUrl - The original external thumbnail URL
-     */
-    getBrowseImageUrl(externalUrl: string): string {
-        const baseUrl = this.getBaseUrl();
-        const token = this.getCurrentToken();
-        const params = new URLSearchParams({ url: externalUrl });
-        if (token) params.append("token", token);
-        return `${baseUrl}/api/browse/ytmusic/image?${params.toString()}`;
-    }
-
     async getYtMusicMixes(): Promise<{
         mixes: Array<{ playlistId: string; title: string; description: string; thumbnails: Array<{ url: string; width: number }>; count: string | null }>;
     }> {
         return this.get("/browse/ytmusic/mixes");
-    }
-
-    // Recommendations
-    async getRecommendationsForYou(limit = 10) {
-        return this.request<{ artists: ApiData[] }>(
-            `/recommendations/for-you?limit=${limit}`
-        );
-    }
-
-    async getSimilarArtists(seedArtistId: string, limit = 20) {
-        return this.request<{ recommendations: ApiData[] }>(
-            `/recommendations?seedArtistId=${seedArtistId}&limit=${limit}`
-        );
-    }
-
-    async getSimilarAlbums(seedAlbumId: string, limit = 20) {
-        return this.request<{ recommendations: ApiData[] }>(
-            `/recommendations/albums?seedAlbumId=${seedAlbumId}&limit=${limit}`
-        );
-    }
-
-    async getSimilarTracks(seedTrackId: string, limit = 20, artist?: string, title?: string) {
-        const params = new URLSearchParams({ seedTrackId, limit: String(limit) });
-        if (artist) params.set("artist", artist);
-        if (title) params.set("title", title);
-        return this.request<{ recommendations: ApiData[] }>(
-            `/recommendations/tracks?${params.toString()}`
-        );
-    }
-
-    async getFreshPreviewUrl(playlistId: string, pendingTrackId: string) {
-        return this.request<{ previewUrl: string }>(
-            `/playlists/${playlistId}/pending/${pendingTrackId}/preview`
-        );
-    }
-
-    // Play tracking
-    async logPlay(trackRef: AddToPlaylistRef) {
-        return this.request<ApiData>("/plays", {
-            method: "POST",
-            body: JSON.stringify(trackRef),
-        });
-    }
-
-    async getRecentPlays(limit = 50) {
-        return this.request<ApiData[]>(`/plays?limit=${limit}`);
-    }
-
-    async getPlayHistorySummary() {
-        return this.request<{
-            allTime: number;
-            last7Days: number;
-            last30Days: number;
-            last365Days: number;
-        }>("/plays/summary");
-    }
-
-    async clearPlayHistory(range: "7d" | "30d" | "365d" | "all") {
-        return this.request<{
-            success: boolean;
-            range: "7d" | "30d" | "365d" | "all";
-            deletedCount: number;
-        }>(`/plays/history?range=${range}`, {
-            method: "DELETE",
-        });
-    }
-
-    // Settings
-    async getSettings() {
-        return this.request<ApiData>("/settings");
-    }
-
-    async updateSettings(settings: ApiData) {
-        return this.request<ApiData>("/settings", {
-            method: "POST",
-            body: JSON.stringify(settings),
-        });
-    }
-
-    // Profile Picture
-    async uploadProfilePicture(file: File): Promise<{ success: boolean }> {
-        const formData = new FormData();
-        formData.append("file", file);
-        const baseUrl = this.getBaseUrl();
-        const token = this.getCurrentToken();
-        const res = await fetch(`${baseUrl}/api/settings/profile-picture`, {
-            method: "POST",
-            headers: { Authorization: `Bearer ${token}` },
-            body: formData,
-        });
-        if (!res.ok) {
-            const err = await res.json().catch(() => ({ error: "Upload failed" }));
-            throw new Error(err.error || "Upload failed");
-        }
-        return res.json();
-    }
-
-    async deleteProfilePicture(): Promise<{ success: boolean }> {
-        return this.request<{ success: boolean }>("/settings/profile-picture", {
-            method: "DELETE",
-        });
-    }
-
-    getProfilePictureUrl(userId: string): string {
-        const baseUrl = this.getBaseUrl();
-        const token = this.getCurrentToken();
-        const params = token ? `?token=${encodeURIComponent(token)}` : "";
-        return `${baseUrl}/api/social/profile-picture/${encodeURIComponent(userId)}${params}`;
-    }
-
-    // System Features
-    async getFeatures(): Promise<{
-        musicCNN: boolean;
-        vibeEmbeddings: boolean;
-        audioAnalysis: boolean;
-        discovery: boolean;
-        autoPlaylists: boolean;
-    }> {
-        return this.request<{
-            musicCNN: boolean;
-            vibeEmbeddings: boolean;
-            audioAnalysis: boolean;
-            discovery: boolean;
-            autoPlaylists: boolean;
-        }>("/system/features");
-    }
-
-    // System UI Settings (non-sensitive, available to all authenticated users)
-    async getUiSettings(): Promise<{ showVersion: boolean }> {
-        return this.request<{ showVersion: boolean }>("/system/ui-settings");
-    }
-
-    // System Settings
-    async getSystemSettings() {
-        return this.request<ApiData>("/system-settings");
-    }
-
-    async updateSystemSettings(settings: ApiData) {
-        return this.request<ApiData>("/system-settings", {
-            method: "POST",
-            body: JSON.stringify(settings),
-        });
-    }
-
-    async clearAllCaches() {
-        return this.request<ApiData>("/system-settings/clear-caches", {
-            method: "POST",
-        });
-    }
-
-    async cleanupStaleJobs() {
-        return this.request<{
-            success: boolean;
-            cleaned: {
-                discoveryBatches: { cleaned: number; ids: string[] };
-                downloadJobs: { cleaned: number; ids: string[] };
-                spotifyImportJobs: { cleaned: number; ids: string[] };
-                bullQueues: { cleaned: number; queues: string[] };
-            };
-            totalCleaned: number;
-        }>("/settings/cleanup-stale-jobs", {
-            method: "POST",
-        });
-    }
-
-    // System Settings Tests
-    async testLidarr(url: string, apiKey: string) {
-        return this.request<ServiceTestResult>("/system-settings/test-lidarr", {
-            method: "POST",
-            body: JSON.stringify({ url, apiKey }),
-        });
-    }
-
-    async testLastfm(apiKey: string) {
-        return this.request<ServiceTestResult>("/system-settings/test-lastfm", {
-            method: "POST",
-            body: JSON.stringify({ lastfmApiKey: apiKey }),
-        });
-    }
-
-    async testOpenai(apiKey: string, model: string) {
-        return this.request<ServiceTestResult>("/system-settings/test-openai", {
-            method: "POST",
-            body: JSON.stringify({ apiKey, model }),
-        });
-    }
-
-    async testFanart(apiKey: string) {
-        return this.request<ServiceTestResult>("/system-settings/test-fanart", {
-            method: "POST",
-            body: JSON.stringify({ fanartApiKey: apiKey }),
-        });
-    }
-
-    async testAudiobookshelf(url: string, apiKey: string) {
-        return this.request<ServiceTestResult>("/system-settings/test-audiobookshelf", {
-            method: "POST",
-            body: JSON.stringify({ url, apiKey }),
-        });
-    }
-
-    async testSoulseek(username: string, password: string) {
-        return this.request<ServiceTestResult>("/system-settings/test-soulseek", {
-            method: "POST",
-            body: JSON.stringify({ username, password }),
-        });
-    }
-
-    async testSpotify(clientId: string, clientSecret: string) {
-        return this.request<ServiceTestResult>("/system-settings/test-spotify", {
-            method: "POST",
-            body: JSON.stringify({ clientId, clientSecret }),
-        });
-    }
-
-    async testTidal() {
-        return this.request<ServiceTestResult>("/system-settings/test-tidal", {
-            method: "POST",
-        });
-    }
-
-    async tidalDeviceAuth() {
-        return this.request<{
-            device_code: string;
-            user_code: string;
-            verification_uri: string;
-            verification_uri_complete: string;
-            expires_in: number;
-            interval: number;
-        }>("/system-settings/tidal-auth/device", {
-            method: "POST",
-        });
-    }
-
-    async tidalPollAuth(deviceCode: string) {
-        return this.request<{
-            status?: "pending";
-            success?: boolean;
-            user_id?: string;
-            country_code?: string;
-            username?: string;
-        }>("/system-settings/tidal-auth/token", {
-            method: "POST",
-            body: JSON.stringify({ device_code: deviceCode }),
-        });
     }
 
     // Downloads (Lidarr)
@@ -1253,23 +850,6 @@ class ApiClient extends WithPlaylists(WithLibrary(ApiClientCore)) {
         );
     }
 
-    async getTrackPreview(artistName: string, trackTitle: string) {
-        return this.request<{ videoId: string }>(
-            `/artists/preview/${encodeURIComponent(
-                artistName
-            )}/${encodeURIComponent(trackTitle)}`
-        );
-    }
-
-    getPreviewStreamUrl(videoId: string): string {
-        const baseUrl = `${this.getBaseUrl()}/api/artists/preview-stream/${encodeURIComponent(videoId)}`;
-        const token = this.getCurrentToken();
-        if (token) {
-            return `${baseUrl}?token=${encodeURIComponent(token)}`;
-        }
-        return baseUrl;
-    }
-
     // Audiobooks
     async getAudiobooks() {
         return this.request<ApiData[]>("/audiobooks");
@@ -1449,44 +1029,6 @@ class ApiClient extends WithPlaylists(WithLibrary(ApiClientCore)) {
         );
     }
 
-    // Playback State (cross-device sync)
-    async getPlaybackState() {
-        return this.request<ApiData>("/playback-state", {
-            headers: {
-                "X-Playback-Device-Id": this.getPlaybackDeviceId(),
-            },
-        });
-    }
-
-    async savePlaybackState(state: {
-        playbackType: string;
-        trackId?: string;
-        audiobookId?: string;
-        podcastId?: string;
-        queue?: ApiData[];
-        currentIndex?: number;
-        isShuffle?: boolean;
-        isPlaying?: boolean;
-        currentTime?: number;
-    }) {
-        return this.request<ApiData>("/playback-state", {
-            method: "POST",
-            headers: {
-                "X-Playback-Device-Id": this.getPlaybackDeviceId(),
-            },
-            body: JSON.stringify(state),
-        });
-    }
-
-    async clearPlaybackState() {
-        return this.request<void>("/playback-state", {
-            method: "DELETE",
-            headers: {
-                "X-Playback-Device-Id": this.getPlaybackDeviceId(),
-            },
-        });
-    }
-
     // Search
     async search(
         query: string,
@@ -1584,72 +1126,6 @@ class ApiClient extends WithPlaylists(WithLibrary(ApiClientCore)) {
                 title,
             }),
         });
-    }
-
-    // Programmatic Mixes
-    async getMixes() {
-        return this.request<ApiData[]>("/mixes");
-    }
-
-    async getMix(id: string) {
-        return this.request<ApiData>(`/mixes/${id}`);
-    }
-
-    async refreshMixes() {
-        return this.request<{ message: string; mixes: ApiData[] }>(
-            "/mixes/refresh",
-            {
-                method: "POST",
-            }
-        );
-    }
-
-    async saveMixAsPlaylist(mixId: string, customName?: string) {
-        return this.request<{ id: string; name: string; trackCount: number }>(
-            `/mixes/${mixId}/save`,
-            {
-                method: "POST",
-                body: customName
-                    ? JSON.stringify({ name: customName })
-                    : undefined,
-            }
-        );
-    }
-
-    // Mood on Demand (Legacy)
-    async getMoodPresets() {
-        return this.request<MoodPreset[]>("/mixes/mood/presets");
-    }
-
-    async generateMoodMix(params: MoodMixParams) {
-        return this.request<ApiData>("/mixes/mood", {
-            method: "POST",
-            body: JSON.stringify(params),
-        });
-    }
-
-    // New Mood Bucket System (simplified, pre-computed)
-    async getMoodBucketPresets() {
-        return this.request<MoodBucketPreset[]>("/mixes/mood/buckets/presets");
-    }
-
-    async getMoodBucketMix(mood: MoodType) {
-        return this.request<MoodBucketMix>(`/mixes/mood/buckets/${mood}`);
-    }
-
-    async saveMoodBucketMix(mood: MoodType) {
-        return this.request<SavedMoodMixResponse>(
-            `/mixes/mood/buckets/${mood}/save`,
-            { method: "POST" }
-        );
-    }
-
-    async backfillMoodBuckets() {
-        return this.request<{
-            success: boolean;
-            processed: number;
-            assigned: number;
-        }>("/mixes/mood/buckets/backfill", { method: "POST" });
     }
 
     // Enrichment
